Refactor: moved the createUserPreferences to the UserService && moved the call to the createUserPreferences when the connection is updated with a 'accepted' status
